push

+++ test/Console.t.sol
@@ -1,11 +1,11 @@
-// SPDX-License-Identifier: UNLICENSED 
+// SPDX-License-Identifier: UNLICENSED
 pragma solidity 0.8.28;
 
 import "forge-std/Test.sol";
 
 contract ConsoleTest is Test {
     function testLogSomething() public view {
-        console.log("Log something here", uint(123));
+        console.log("Log something here", uint256(123));
 
         int256 x = -1;
         console.logInt(x);
